fix(auth): share bearer token parsing
