+++ scripts/benchmarks/user-events-transform-benchmark.mjs
@@ -28,7 +28,7 @@ function transformUserEventCounts(data, timeField = 'event_time') {
 
     userSet.add(user)
     if (acc[event_time] === undefined) acc[event_time] = {}
-    acc[event_time][user] = count
+    acc[event_time][user] = (acc[event_time][user] ?? 0) + count
     return acc
   }, {})
